Rename existingThoughtMove -> moveThought
