+++ src/action-creators/index.ts
@@ -36,7 +36,7 @@ export const expandHoverTop = reducerToThunk<typeof reducers.expandHoverTop>('ex
 export const expandBottom = reducerToThunk<typeof reducers.expandBottom>('expandBottom')
 export const editThought = reducerToThunk<typeof reducers.editThought>('editThought')
 export const existingThoughtDelete = reducerToThunk<typeof reducers.existingThoughtDelete>('existingThoughtDelete')
-export const existingThoughtMove = reducerToThunk<typeof reducers.existingThoughtMove>('existingThoughtMove')
+export const moveThought = reducerToThunk<typeof reducers.moveThought>('moveThought')
 export const extractThought = reducerToThunk<typeof reducers.extractThought>('extractThought')
 export const indent = reducerToThunk<typeof reducers.indent>('indent')
 export const isPushing = reducerToThunk<typeof reducers.isPushing>('isPushing')
