+++ src/reducers/__tests__/moveThought.ts
@@ -1,14 +1,14 @@
 import { HOME_TOKEN } from '../../constants'
 import { equalArrays, initialState, reducerFlow } from '../../util'
 import { exportContext, getContexts, getThought, getAllChildren, getChildrenRanked, getRankAfter, getChildren, isPending } from '../../selectors'
-import { existingThoughtMove, importText, newSubthought, newThought, setCursor } from '../../reducers'
+import { moveThought, importText, newSubthought, newThought, setCursor } from '../../reducers'
 import checkDataIntegrity from '../../test-helpers/checkDataIntegrity'
 import { State } from '../../util/initialState'
 import { SimplePath } from '../../types'
 import { store as appStore } from '../../store'
 import testTimer from '../../test-helpers/testTimer'
 import { initialize } from '../../initialize'
-import { clear, importText as importTextAction, existingThoughtMove as existingThoughtMoveAction } from '../../action-creators'
+import { clear, importText as importTextAction, moveThought as existingThoughtMoveAction } from '../../action-creators'
 import { setCursorFirstMatchActionCreator } from '../../test-helpers/setCursorFirstMatch'
 
 const timer = testTimer()
@@ -18,7 +18,7 @@ it('move within root', () => {
   const steps = [
     newThought('a'),
     newThought('b'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'b', rank: -1 }],
     }),
@@ -54,7 +54,7 @@ it('persist id on move', () => {
   const oldId = oldExactThought?.id
 
   const steps2 = [
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'a', rank: 0 }, { value: 'a1', rank: 0 }],
       newPath: [{ value: 'a1', rank: 1 }],
     }),
@@ -73,7 +73,7 @@ it('move within context', () => {
     newThought('a'),
     newSubthought('a1'),
     newThought('a2'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'a', rank: 0 }, { value: 'a2', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'a2', rank: -1 }],
     }),
@@ -104,7 +104,7 @@ it('move across contexts', () => {
     newSubthought('a1'),
     newThought({ value: 'b', at: [{ value: 'a', rank: 0 }] }),
     newSubthought('b1'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 0 }, { value: 'b1', rank: 0 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'b1', rank: 1 }],
     }),
@@ -138,7 +138,7 @@ it('move descendants', () => {
     newThought({ value: 'b', at: [{ value: 'a', rank: 0 }] }),
     newSubthought('b1'),
     newSubthought('b1.1'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'b', rank: -1 }],
     }),
@@ -183,7 +183,7 @@ it('moving cursor thought should update cursor', () => {
     newThought('a'),
     newSubthought('a1'),
     newThought('a2'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'a', rank: 0 }, { value: 'a2', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'a2', rank: -1 }],
     }),
@@ -204,7 +204,7 @@ it('moving ancestor of cursor should update cursor', () => {
     newThought('b'),
     newSubthought('b1'),
     newSubthought('b1.1'),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'b', rank: -1 }],
     }),
@@ -227,7 +227,7 @@ it('moving unrelated thought should not update cursor', () => {
     newSubthought('b1'),
     newSubthought('b1.1'),
     setCursor({ path: [{ value: 'a', rank: 0 }] }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'b', rank: -1 }],
     }),
@@ -252,7 +252,7 @@ it('move root thought into another root thought', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'a', rank: 1 }],
       newPath: [{ value: 'x', rank: 0 }, { value: 'a', rank: 0 }],
     }),
@@ -297,7 +297,7 @@ it('move descendants with siblings', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'a', rank: 0 }, { value: 'b', rank: 0 }],
       newPath: [{ value: 'b', rank: 1 }],
     }),
@@ -344,7 +344,7 @@ it('merge duplicate with new rank', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'm', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'm', rank: 0 }],
     }),
@@ -384,7 +384,7 @@ it('merge with duplicate with duplicate rank', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'm', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'm', rank: 0 }],
     }),
@@ -424,7 +424,7 @@ it('move with duplicate descendant', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'b', rank: 0 }],
     }),
@@ -461,7 +461,7 @@ it('move with hash matched descendant', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'b', rank: 1 }],
       newPath: [{ value: 'a', rank: 0 }, { value: 'b', rank: 0 }],
     }),
@@ -496,7 +496,7 @@ it('move with nested duplicate thoughts', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'c', rank: 1 }, { value: 'a', rank: 0 }],
       newPath: [{ value: 'a', rank: 0 }],
     }),
@@ -534,7 +534,7 @@ it('move with nested duplicate thoughts and merge their children', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'p', rank: 1 }, { value: 'a', rank: 0 }],
       newPath: [{ value: 'a', rank: 0 }],
     }),
@@ -583,7 +583,7 @@ it('data integrity test', () => {
 
   const steps = [
     importText({ text }),
-    existingThoughtMove({
+    moveThought({
       oldPath: [{ value: 'k', rank: 0 }, { value: 'a', rank: 0 }],
       newPath: [{ value: 'm', rank: 0 }],
     }),
@@ -610,7 +610,7 @@ it('consitent rank between thoughtIndex and contextIndex on duplicate merge', ()
   const steps = [
     importText({ text }),
     (state: State) =>
-      existingThoughtMove(state, {
+      moveThought(state, {
         oldPath: [{ value: 'a', rank: 0 }, { value: 'b', rank: 0 }],
         // Note: Here new rank will be 0.5 because it's calculated between a (0) and b (1)
         newPath: [{ value: 'b', rank: getRankAfter(state, [{ value: 'a', rank: 0 }] as SimplePath) }],
