+++ src/reducers/bumpThoughtDown.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash'
-import { editThought, existingThoughtMove, newThoughtSubmit, setCursor, subCategorizeOne, editableRender } from '../reducers'
+import { editThought, moveThought, newThoughtSubmit, setCursor, subCategorizeOne, editableRender } from '../reducers'
 import { getPrevRank, getRankBefore, getAllChildren, simplifyPath, rootedParentOf } from '../selectors'
 import { parentOf, headValue, pathToContext, reducerFlow, unroot } from '../util'
 import { State } from '../util/initialState'
@@ -33,7 +33,7 @@ const bumpThoughtDown = (state: State, { simplePath }: { simplePath?: SimplePath
   return reducerFlow([
 
     // modify the rank to get the thought to re-render (via the Subthoughts child key)
-    existingThoughtMove({
+    moveThought({
       oldPath: simplePath,
       newPath: simplePathWithNewRank,
     }),
