+++ src/reducers/index.ts
@@ -30,7 +30,7 @@ export { default as expandHoverTop } from './expandHoverTop'
 export { default as expandBottom } from './expandBottom'
 export { default as editThought } from './editThought'
 export { default as existingThoughtDelete } from './existingThoughtDelete'
-export { default as existingThoughtMove } from './existingThoughtMove'
+export { default as moveThought } from './moveThought'
 export { default as expandContextThought } from './expandContextThought'
 export { default as extractThought } from './extractThought'
 export { default as fontSize } from './fontSize'
